Addition of perl-modules that perl-www-mechanize depend on.
WWW::Mechanize depends on a number of these modules, which have
their own dependencies also. Updated the dependencies for libwww
and perl-www-mechanize to include the required modules for an
HTTP connection.

Also included the perl-devel-modlist module as it was useful when
debugging what dependencies a module had and if there were
version conflicts